Implement Moves and Undoing Moves
Cars are moved by repeatedly moving the field representing
the end of the car in front of the first field. If the move
step is negative, this of course works in reverse.

With this way of moving cars, the needed number of writes is minimized,
which should give a speed boost for especially large cars.

Undoing a move is done by first deducing the orientation of the moved
car, and then moving back the original spot in reverse.