Parser Rework: Change Board FEN Format
The format of the FEN is now changed to include dimensions
explicitly, as well as information about the board, such as
relative difficulty to other boards and minimum number of moves
to solve. Every field is separated by a ":" to simplify parsing.

Since the solution and difficulty of a board is not always known,
a "?" is also accepted for these fields to signify an optional null value.

The actual board also doesn't contain newlines anymore, since they
caused problems when parsing and became unnecessary with the
introduction of explicit dimensions.