Harden Parser with errors for car size and orientation functions
These functions can now return an error specifying empty fields,
out of bounds positions and invalid moves.

This commit also adds a different mechanismn for determining the
goal field. Instead of a single position, we now store the goal
lane as well as an orientation. This approach gets rid of the
problems that happened in connection with bad offset_position()
calls for horizontal goals.