Board: Implement isMoveToSecluded heuristic
This heuristic checks if the previous move positioned the
moving car at a location no other car could move to.

To efficiently implement this, we just need to shoot rays
in all 4 directions starting at the destination square
and checking if the orientation of any hit car matches
the one of the ray.