Board: Implement initialBoardDistance heuristic
This heuristic essentially calculates the sum of the
differences in the positions of all cars to the
initial board.

Since cars can't swap places, we can easily iterate
through all rows and columns and add the differences
between representative car positions together.

Another approach would be to calculate this through
the move history, by checking how much each car was
moved in total. But this would mean making the move
history available as context in the brain, which
would maybe make things more complicated to implement.