Board: Implement BlockersLowerBound heuristic
This heuristic computes an approximation of how many
moves are needed for the goal car to get to the exit.

This is done by recursively identifying and moving
blocking cars, while keeping a cache of already
calculated moves to prevent cycles.

Note that this heuristic may sometimes be incorrect,
since no car is actually ever moved and every move
is looked at in isolation, but it should still
provide a good estimate of how far the current board
is from a solution.